chore: rename packages to align with feedback from other PR
