+++ internal/hooks/hookshttp/hookshttp_test.go
@@ -1,4 +1,4 @@
-package v0http
+package hookshttp
 
 import (
 	"bytes"
