+++ internal/hooks/hookspgfunc/hookspgfunc.go
@@ -1,4 +1,4 @@
-package v0pgfunc
+package hookspgfunc
 
 import (
 	"context"
