+++ internal/hooks/hookspgfunc/hookspgfunc_test.go
@@ -1,4 +1,4 @@
-package v0pgfunc
+package hookspgfunc
 
 import (
 	"context"
